Add an example where inplace is used with xargs
So I came across this looking for inplace editing with jq. My desired use case was to find all json files with a certain field, and set that field to '[]' if it existed.

I thought it'd serve as useful documentation to add an example of using inplace with xargs, although this example is probably a bit convoluted, so please tidy it up if you thing you've got a better version.